bean-validation-demo/src/main/java/no/nith/pg5100/Customer.java
package no.nith.pg5100;

import org.hibernate.validator.constraints.CreditCardNumber;

import javax.validation.constraints.AssertTrue;
import javax.validation.constraints.Min;
import javax.validation.constraints.NotNull;
import javax.validation.constraints.Pattern;
import java.time.LocalDate;

@CustomerCreditCard
public class Customer {
    @Min(10)
    private int id;
    @NotNull
    private String username;
    @NotNull
    @Pattern(regexp = "^\\S+@\\S+\\.\\S+$")
    private String email;
    @AssertTrue
    private boolean active;

    private String cardNo;
    private LocalDate expirationDate;

    public int getId() {
        return id;
    }

    public void setId(int id) {
        this.id = id;
    }

    public String getUsername() {
        return username;
    }

    public void setUsername(String username) {
        this.username = username;
    }

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }

    public boolean isActive() {
        return active;
    }

    public void setActive(boolean active) {
        this.active = active;
    }

    public String getCardNo() {
        return cardNo;
    }

    public void setCardNo(String cardNo) {
        this.cardNo = cardNo;
    }

    public LocalDate getExpirationDate() {
        return expirationDate;
    }

    public void setExpirationDate(LocalDate expirationDate) {
        this.expirationDate = expirationDate;
    }
}
